성적 유효성 체크 완료

+++ src/main/java/com/green/university/dto/PenaltyResultDto.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @AllArgsConstructor
 public class PenaltyResultDto {
     private long totalAbsent;
-    private int penalty;
+    private double penalty;
 }
